requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
facenet-pytorch==2.6.0
filelock==3.18.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
pillow==10.2.0
requests==2.32.3
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
